pyproject.toml
[project]
name = "fusionpipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Francesco Carpanese", email = "francescocarpanese@hotmail.it" }
]
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.29.1",
    "networkx>=3.5",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "toml>=0.10.2",
    "uvicorn>=0.34.2",
    "nbconvert>=7.16.6",
    "ray[default]>=2.48.0",
    "ipython>=9.3.0",
    "fp-user-utils",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = [
    "fp_user_utils",
]

[tool.uv.sources]
fp-user-utils = { workspace = true }
